core: Split token from governance functionality
